refactor(email-login): inputText부분 map함수로 관리
